changed to int32 for inputs

+++ lstm_tf_imdb.py
@@ -83,7 +83,7 @@ class LSTM_Model(object):
         # learning rate as a tf variable. Its value is therefore session dependent
         self._lr = tf.Variable(config.learning_rate, trainable=False)
 
-        self._inputs = tf.placeholder(tf.int64,[config.MAXLEN,BATCH_SIZE],name='embedded_inputs')
+        self._inputs = tf.placeholder(tf.int32,[config.MAXLEN,BATCH_SIZE],name='embedded_inputs')
         self._targets = tf.placeholder(tf.float32, [None, 2],name='targets')
         self._mask = tf.placeholder(tf.float32, [None, None],name='mask')
